Support custom LeetCode types

+++ libcaide/tests/topcoder/inlined/prob1/submission.cpp
@@ -12,7 +12,7 @@
 class ClassName {
 public:
     long long methodName(int i, double d) {
-        long long result = 0;
+        long long result = (long long)0;
         return result;
     }
 };
